requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlmodel>=0.0.14
sentence-transformers>=2.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
groq>=0.4.0
openai>=1.12.0
jinja2>=3.1.3
argon2-cffi>=23.1.0
pytest>=7.0.0
httpx>=0.24.0
pytest-asyncio>=0.21.0